Added throwOnFail
DoesCIExist returns a boolean for existence. This is suitable for a conditional check. Added the throwOnFail - if a False is returned then this task will Fail. This allows the Template creator to use this task as a hard stop, rather than as a conditional or needing a second task to confirm success later.